+++ src/FlaUILibrary/keywords/window.py
@@ -11,7 +11,7 @@ class WindowKeywords:
 
     def __init__(self, container: AutomationInterfaceContainer):
         """
-        Constructor for mouse keywords.
+        Constructor for window keywords.
 
         ``container`` User automation container to handle element interaction.
         """
@@ -30,10 +30,31 @@ class WindowKeywords:
         Example:
         | Launch Application  <APPLICATION>           |
         | Close Window  <XPATH_TO_APPLICATION_WINDOW> |
-
         """
         module = self._container.create_or_get_module()
         element = module.get_element(identifier, InterfaceType.WINDOW, msg=msg)
         module.action(Window.Action.CLOSE_WINDOW,
                       Window.Container(element=element),
+                      msg)
+
+    @keyword
+    def resize_window(self, identifier, width: int, height: int, msg=None):
+        """
+        Try to resize window from element to given width and height.
+
+        Arguments:
+        | Argument   | Type   | Description                             |
+        | identifier | string | XPath identifier from element to search |
+        | width      | int    | New width of the window                  |
+        | height     | int    | New height of the window                 |
+        | msg        | string | Custom error message                     |
+
+        Example:
+        | Launch Application  <APPLICATION>                  |
+        | Resize Window  <XPATH_TO_APP_WINDOW>  1024  768     |
+        """
+        module = self._container.create_or_get_module()
+        element = module.get_element(identifier, InterfaceType.WINDOW, msg=msg)
+        module.action(Window.Action.RESIZE_WINDOW,
+                      Window.Container(element=element, width=width, height=height),
                       msg)